+++ app/src/main/java/com/example/namo/data/dao/EventDao.kt
@@ -35,7 +35,7 @@ interface EventDao {
     fun getNotUploadedEvent() : List<Event>
 
     @Query("UPDATE calendar_event_table SET event_upload=:isUpload, event_server_idx=:serverIdx, event_state=:state WHERE eventId=:eventId")
-    fun updateEventAfterUpload(eventId : Long, isUpload : Int, serverIdx : Int, state : String)
+    fun updateEventAfterUpload(eventId : Long, isUpload : Int, serverIdx : Long, state : String)
 
     @Query("DELETE FROM calendar_event_table WHERE eventId=:eventId")
     fun deleteEventById(eventId : Long)
